# packages/game/infra/api-spec.yaml
openapi: 3.0.0
info:
  title: Game data API for grid-wolf
  description: |-
    This API provides endpoints for updating game data for users of grid-wolf.
  version: 0.1.0
security: 
  - api_key: []
paths:
  '/game':
    put:
      description: Upload a new or replacement game entry
      operationId: putGame
      requestBody: 
        $ref: '#/components/requestBodies/PutGameRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad request
      x-amazon-apigateway-integration:
        type: aws_proxy
        uri: arn:aws:apigateway:{{region}}:lambda:path/2015-03-31/functions/{{handler}}/invocations
        httpMethod: POST
  '/game/{gameId}':
    get:
      description: Upload a new or replacement game entry
      operationId: getGame
      parameters: 
        - in: path
          name: gameId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
        - api_key: []
      x-amazon-apigateway-integration:
        type: aws_proxy
        uri: arn:aws:apigateway:{{region}}:lambda:path/2015-03-31/functions/{{handler}}/invocations
        httpMethod: POST
  '/games':
    get:
      description: Upload a new or replacement game entry
      operationId: getGames
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
        - api_key: []
      x-amazon-apigateway-integration:
        type: aws_proxy
        uri: arn:aws:apigateway:{{region}}:lambda:path/2015-03-31/functions/{{handler}}/invocations
        httpMethod: POST
components:
  schemas:
    Game:
      type: object
      required:
        - id
        - userId
        - name
        - timestamp
      properties:
        id:
          type: string
        userId:
          type: string
        name:
          type: string
        timestamp:
          type: string
    ApiResponse:
      type: object
      properties:
        statusCode:
          type: integer
        message:
          type: string
  requestBodies:
    PutGameRequest:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Game'
  securitySchemes:
    api_key:
      type: apiKey
      name: x-api-key
      in: header